move GenSampleFn to util package
