+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/util/CvrBuilder.kt
@@ -3,7 +3,6 @@ package org.cryptobiotic.rlauxe.util
 import org.cryptobiotic.rlauxe.core.AuditContest
 import org.cryptobiotic.rlauxe.core.Cvr
 import org.cryptobiotic.rlauxe.core.SocialChoiceFunction
-import org.cryptobiotic.rlauxe.core.flipExactVotes
 import kotlin.random.Random
 
 // for testing, add here to share between modules
