+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/util/GenSampleFn.kt
@@ -1,5 +1,8 @@
-package org.cryptobiotic.rlauxe.core
+package org.cryptobiotic.rlauxe.util
 
+import org.cryptobiotic.rlauxe.core.AssorterFunction
+import org.cryptobiotic.rlauxe.core.ComparisonAssorter
+import org.cryptobiotic.rlauxe.core.Cvr
 import kotlin.math.ln
 import kotlin.random.Random
 
@@ -91,7 +94,8 @@ class ComparisonNoErrors(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter
 
 // generate mvr by starting with cvrs and flipping exact # votes (type 2 errors only)
 // to make mvrs have mvrMean. TODO withReplacement only?
-data class ComparisonWithErrors(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter, val mvrMean: Double): GenSampleFn {
+data class ComparisonWithErrors(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter, val mvrMean: Double):
+    GenSampleFn {
     val N = cvrs.size
     val mvrs : List<Cvr>
     val permutedIndex = MutableList(N) { it }
