+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestAlphaMartComparison.kt
@@ -2,6 +2,8 @@ package org.cryptobiotic.rlauxe.core
 
 import org.cryptobiotic.rlauxe.doublePrecision
 import org.cryptobiotic.rlauxe.makeStandardComparisonAssorter
+import org.cryptobiotic.rlauxe.util.ComparisonNoErrors
+import org.cryptobiotic.rlauxe.util.GenSampleFn
 import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
 import kotlin.math.max
 import kotlin.test.Test
