+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestWelford.kt
@@ -1,6 +1,7 @@
 package org.cryptobiotic.rlauxe.core
 
 import org.cryptobiotic.rlauxe.doublePrecision
+import org.cryptobiotic.rlauxe.util.Bernoulli
 import kotlin.math.sqrt
 import kotlin.test.Test
 import kotlin.test.assertEquals
