+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/comparison/CompareAuditType.kt
@@ -16,9 +16,9 @@ import kotlinx.coroutines.sync.withLock
 
 import kotlinx.coroutines.yield
 import org.cryptobiotic.rlauxe.core.AuditContest
-import org.cryptobiotic.rlauxe.core.ComparisonNoErrors
+import org.cryptobiotic.rlauxe.util.ComparisonNoErrors
 import org.cryptobiotic.rlauxe.core.Cvr
-import org.cryptobiotic.rlauxe.core.PollWithoutReplacement
+import org.cryptobiotic.rlauxe.util.PollWithoutReplacement
 import org.cryptobiotic.rlauxe.core.makeComparisonAudit
 import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
 import org.cryptobiotic.rlauxe.core.makePollingAudit
