+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/sim/BettingTask.kt
@@ -2,6 +2,8 @@ package org.cryptobiotic.rlauxe.sim
 
 import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.makeStandardComparisonAssorter
+import org.cryptobiotic.rlauxe.util.ComparisonWithErrorRates
+import org.cryptobiotic.rlauxe.util.GenSampleFn
 
 data class BettingTask(
     val idx: Int,
