+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/sim/PollingTask.kt
@@ -2,6 +2,8 @@ package org.cryptobiotic.rlauxe.sim
 
 import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.makeStandardPluralityAssorter
+import org.cryptobiotic.rlauxe.util.GenSampleFn
+import org.cryptobiotic.rlauxe.util.PollWithoutReplacement
 import kotlin.math.max
 
 data class PollingTask(
